Implement CSV export as plugin
