+++ src/Persistence/CMakeLists.txt
@@ -60,10 +60,7 @@ target_sources(Persistence
         src/Service/DatabaseService.h
         src/Service/CSVService.cpp
         src/Service/CSVService.h
-        src/CSVConst.cpp
         src/CSVConst.h
-        src/Export/CSVExport.cpp
-        src/Export/CSVExport.h
         src/Import/CSVImport.cpp
         src/Import/CSVImport.h
 )
