+++ src/Plugin/src/Export/CSVExport/src/CSVExportPlugin.json
@@ -1,4 +1,4 @@
 {
     "uuid": "0fc478c9-97f8-4b4b-bea7-0e53833ab6aa",
-    "name": "CSV Export"
+    "name": "Comma-Separated Values (CSV)"
 }
