+++ src/UserInterface/src/MainWindow.ui
@@ -239,7 +239,7 @@
      <x>0</x>
      <y>0</y>
      <width>824</width>
-     <height>22</height>
+     <height>21</height>
     </rect>
    </property>
    <widget class="QMenu" name="fileMenu">
@@ -258,7 +258,6 @@
     <addaction name="optimiseLogbookAction"/>
     <addaction name="separator"/>
     <addaction name="importCSVAction"/>
-    <addaction name="exportCSVAction"/>
     <addaction name="exportMenu"/>
     <addaction name="separator"/>
     <addaction name="showSettingsAction"/>
